--- src/bin/abci_node/data/leviathan_db/CURRENT
@@ -1,1 +1,0 @@
-MANIFEST-000005
