--- src/bin/abci_node/data/leviathan_db/IDENTITY
@@ -1,1 +1,0 @@
-d5711101-a2fb-441e-b3e2-20d0ccec95e9
